+++ illumio/util/constants.py
@@ -24,6 +24,8 @@ AMS = 'ams'
 #: Name of the default global IP list.
 ANY_IP_LIST_NAME = 'Any (0.0.0.0/0 and ::/0)'
 
+PORT_MAX = 65535
+
 FQDN_REGEX = re.compile('(?=^.{4,253}$)(^((?!-)[a-zA-Z0-9-]{1,63}(?<!-)\.)+[a-zA-Z]{2,63}$)')
 HREF_REGEX = re.compile('^\/orgs\/\d+\/(?:sec_policy\/(?:active|draft)\/)?(?P<type>[a-zA-Z_]+)\/(?P<uid>[a-zA-Z0-9-]+)$')